clang warning

+++ lib/bytes/src/bytes.cpp
@@ -83,7 +83,8 @@ operator<<(std::ostream& out, const bytes& data)
     return out << to_hex(data);
   }
 
-  bytes abbrev(data.begin(), data.begin() + static_cast<size_t>(threshold));
+  // NOLINTNEXTLINE (cppcoreguidelines-narrowing-conversion)
+  bytes abbrev(data.begin(), data.begin() + threshold);
   return out << to_hex(abbrev) << "...";
 }
